🚚 Unrelated: Move component to correct folder
