+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/TabPayLaterMessaging.js
@@ -1,4 +1,4 @@
-import { PayLaterMessagingHooks } from '../../../data';
+import { PayLaterMessagingHooks } from '../../../../data';
 import { useEffect } from '@wordpress/element';
 
 const TabPayLaterMessaging = () => {
